requirements.txt
Django>=5.0,<5.3
djangorestframework>=3.15,<3.16
gunicorn>=21.2,<23
psycopg2-binary>=2.9,<3.0
python-dotenv>=1.0,<2.0
whitenoise>=6.6,<7
Pillow>=10.0,<12
